+++ CHANGELOG.md
@@ -13,3 +13,9 @@
 ### 🛠 Technical
 
 - Minor bug fixes on tools (CI, csproj, etc...)
+
+## 202
+
+### 🐛 Bug Fixes
+
+- Fixed a bug where the entire query string was encoded instead of only the query parameter keys and values.